feat: adiciona tratamento de erro para quando usuario tenta deletar elemento que esta atrelado a outra tabela

+++ backend/backend/src/main/java/com/repairio/backend/service/CategoriaService.java
@@ -1,17 +1,23 @@
 package com.repairio.backend.service;
 
-import com.repairio.backend.dao.CategoriaDao;
-import com.repairio.backend.model.Categoria;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.repairio.backend.dao.CategoriaDao;
+import com.repairio.backend.dao.SolicitacaoDao;
+import com.repairio.backend.exception.CategoriaConstraintException;
+import com.repairio.backend.model.Categoria;
 
 @Service
 public class CategoriaService {
-    private final CategoriaDao categoriaDao;
 
-    public CategoriaService(CategoriaDao categoriaDao) {
+    private final CategoriaDao categoriaDao;
+    private final SolicitacaoDao solicitacaoDao;
+
+    public CategoriaService(CategoriaDao categoriaDao, SolicitacaoDao solicitacaoDao) {
         this.categoriaDao = categoriaDao;
+        this.solicitacaoDao = solicitacaoDao;
     }
 
     public List<Categoria> findAll() {
@@ -31,6 +37,9 @@ public class CategoriaService {
     }
 
     public void delete(Long id) {
+        if (solicitacaoDao.findByCategoryId(id) != null) {
+            throw new CategoriaConstraintException("Esta categoria está atrelada a uma ou mais solicitações.");
+        }
         categoriaDao.delete(id);
     }
 }
